+++ src/commands.ts
@@ -44,6 +44,7 @@ import {
   getSomeBlockNumberBefore,
   getTypeAndVersion,
   lazyCached,
+  parseExtraArgs,
   parseWithFragment,
   recursiveParseError,
 } from './lib/index.js'
@@ -581,6 +582,12 @@ export function parseBytes({ data, selector }: { data: string; selector?: string
     parsed = parseWithFragment(selector, data)
   } else {
     if (isBytesLike(data)) {
+      const extraArgs = parseExtraArgs(data)
+      if (extraArgs) {
+        const { _tag, ...rest } = extraArgs
+        console.info(`${_tag}:`, rest)
+        return
+      }
       parsed = parseWithFragment(dataSlice(data, 0, 4), dataSlice(data, 4))
     }
     if (!parsed) {
@@ -592,7 +599,18 @@ export function parseBytes({ data, selector }: { data: string; selector?: string
   const name = fragment.constructor.name.replace(/Fragment$/, '')
   console.info(`${name}: ${contract.replace(/_\d\.\d.*$/, '')} ${fragment.format('full')}`)
   if (args) {
-    const formatted = formatResult(args)
+    const formatted = formatResult(args, (val, key) => {
+      if (key === 'extraArgs' && isHexString(val)) {
+        const extraArgs = parseExtraArgs(val)
+        if (extraArgs) {
+          const { _tag, ...rest } = extraArgs
+          return `${_tag}(${Object.entries(rest)
+            .map(([k, v]) => `${k}=${v}`)
+            .join(', ')})`
+        }
+      }
+      return val
+    })
     const ps: unknown[] = []
     if (fragment.name === 'ReceiverError' && args.err === '0x') {
       ps.push('[possibly out-of-gas or abi.decode error]')