+++ src/lib/offchain.ts
@@ -164,7 +164,8 @@ async function getLbtcTokenData(
 ): Promise<(string | undefined)[]> {
   const lbtcDepositHashes = allLogsInRequest
     .filter(({ topics }) => topics[0] === LBTC_EVENT.topicHash)
-    .map(({ topics }) => topics[2])
+    .map(({ topics }) => topics[3])
+  console.log(lbtcDepositHashes);
   if (lbtcDepositHashes.length === 0) return message.tokenAmounts.map(() => undefined)
   return Promise.all(
     message.tokenAmounts.map(async (_, idx) => {
@@ -174,8 +175,8 @@ async function getLbtcTokenData(
       if (destTokenData.length === 66 && lbtcDepositHashes.includes(destTokenData)) {
         try {
           return await getLbtcAttestation(destTokenData, isTestnet)
-        } catch (_) {
-          // fallback: undefined
+        } catch (e) {
+          console.log(e);
         }
       }
     }),