+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+- Allow `parseBytes` command to parse EVMExtraArgs bytearrays, both standalone and in structs (#7)
+
 ## [0.1.2] - 2024-11-25
 - Add public `recursiveParseErrors` function to lib, to return nested/inner ABI errors
 - Use it everywhere: `parseBytes` command, thrown exceptions, and `prettyReceipt` output of failed execs